minimize use of KryoWrapper
