+++ spark/src/main/scala/geotrellis/spark/io/s3/RasterRDDReader.scala
@@ -27,7 +27,6 @@ class RasterRDDReader[K: SpatialComponent: Boundable: AvroRecordCodec: JsonForma
     val keyBounds = getLayerKeyBounds[K](id)
     val keyIndex  = getLayerKeyIndex[K](id)
 
-
     val bucket = metadata.bucket
     val prefix = metadata.key
 
@@ -38,22 +37,19 @@ class RasterRDDReader[K: SpatialComponent: Boundable: AvroRecordCodec: JsonForma
 
     logger.debug(s"Loading layer from $bucket $prefix, ${ranges.length} ranges split into ${bins.length} bins")
 
-    val indexToPath: (Long, Int) => String = encodeIndex
     val writerSchema: Schema = (new Schema.Parser).parse(attributeStore.read[JsObject](id, "schema").toString())
     val maxWidth = maxIndexWidth(keyIndex.toIndex(keyBounds.maxKey))
+    val recordCodec = KeyValueRecordCodec[K, Tile]
+    val boundable = implicitly[Boundable[K]]
+    val includeKey = (key: K) => KeyBounds.includeKey(queryKeyBounds, key)(boundable)
+    val toPath = (index: Long) => encodeIndex(index, maxWidth)
 
-    val BC = KryoWrapper((
-      getS3Client,
-      { (index: Long) => indexToPath(index, maxWidth) },
-      { (key: K) => queryKeyBounds.includeKey(key) },
-      KeyValueRecordCodec[K, Tile],
-      writerSchema
-      ))
+    val BC = KryoWrapper((getS3Client, recordCodec, writerSchema))
 
     val rdd =
       sc.parallelize(bins, bins.size)
         .mapPartitions { partition: Iterator[Seq[(Long, Long)]] =>
-          val (getS3Client, toPath, includeKey, recCodec, schema) = BC.value
+          val (getS3Client, recCodec, schema) = BC.value
           val s3client = getS3Client()
 
           val tileSeq: Iterator[Seq[(K, Tile)]] =
